+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/ShapeSpell.java
@@ -60,7 +60,7 @@ public class ShapeSpell extends BrushSpell
 
 		int maxDimension = parameters.getInt("max_dimension", DEFAULT_MAX_DIMENSION);
 		maxDimension = parameters.getInt("md", maxDimension);
-		maxDimension = (int)(mage.getConstructionMultiplier() * (float)maxDimension);
+		maxDimension = (int)(mage.getConstructionMultiplier() * maxDimension);
 
 		int diameter = radius * 2;
 		if (diameter > maxDimension)
